intervention image

+++ app/Http/Controllers/BooksController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Books;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Laravel\Facades\Image;
+
 
 class BooksController extends Controller
 {
@@ -38,14 +41,30 @@ class BooksController extends Controller
             'judul' => 'required|string',
             'penulis' => 'required|string',
             'harga' => 'required|numeric',
-            'tgl_terbit' => 'required|date'
+            'tgl_terbit' => 'required|date',
+            'cover' => 'image|required|max:1999',
         ]);
+
+        if ($request->hasFile('cover')) {
+            $image = Image::read($request->file('cover'));
+            $filenameWithExt = $request->file('cover')->getClientOriginalName();
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('cover')->getClientOriginalExtension();
+            $basename = uniqid() . time();
+            $smallFilename = "small_{$basename}.{$extension}";
+            $mediumFilename = "medium_{$basename}.{$extension}";
+            $largeFilename = "large_{$basename}.{$extension}";
+            $filenameSimpan = "{$basename}.{$extension}";
+            $path = $request->file('cover')->storeAs('cover_book_images', $filenameSimpan);
+            $image->cover(300, 300, 'center')->save(storage_path("app/public/cover_book_images_resized/{$filenameSimpan}"));
+        }
 
         $buku = new Books();
         $buku->judul = $request->judul; // lihat name di view
         $buku->penulis = $request->penulis;
         $buku->harga = $request->harga;
         $buku->tgl_terbit = $request->tgl_terbit;
+        $buku->photo = $filenameSimpan;
         $buku->save();
 
         return redirect('/books')
@@ -55,9 +74,10 @@ class BooksController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function showCover(string $id)
     {
-        //
+        $books_data = Books::find($id);
+        return view('book.detail', compact('books_data'));
     }
 
     /**
@@ -92,7 +112,13 @@ class BooksController extends Controller
     {
         $buku = Books::find($id);
 
-        $buku ->delete();
+        if($buku){
+
+            Storage::disk('public')->delete('cover_book_images_resized/' . $buku->photo);
+            Storage::disk('public')->delete('cover_book_images/' . $buku->photo);
+
+            $buku ->delete();
+        }
 
         return redirect('/books')->with('pesan', 'Data buku berhasil di Hapus');
     }
